pyproject.toml
[project]
name = "audit"
version = "0.1.0"
description = "US manufacturing trade analysis with China dependency and defense criticality metrics"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.0",
    "fastapi>=0.119.0",
    "ipykernel>=6.30.1",
    "jinja2>=3.1.6",
    "openai>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "ruff>=0.14.0",
    "uvicorn>=0.37.0",
]
